Bill S7715A version original - Jeremy Cooney
BILL NUMBER: S7715 SPONSOR: COONEY TITLE OF BILL: An act to amend the criminal procedure law, in relation to criminal possession of stolen property SUMMARY OF SPECIFIC PROVISIONS: Section 1 amends section 510.10 of the criminal procedure law by adding the crimes of criminal possession of stolen property in the fourth, third, second, and first to the list of qualifying offenses. Section 2 amends section 530.20 of the criminal procedure law to make corresponding changes.