Bill S7711 version original - Robert Ortt
BILL NUMBER: S7711REVISED 4/30/24 SPONSOR: ORTT TITLE OF BILL: An act to amend the education law, in relation to adopting the PA licen- sure compact PURPOSE OR GENERAL IDEA OF BILL: Enacts the PA licensure compact to strengthen medical services that are provided throughout the state. SUMMARY OF PROVISIONS: Section 1. This act enacts into law components of legislation relating to enacting the interstate PA compact. The education law is amended by adding a new section, 6546-a.