Bill S7722 version original - Peter Oberacker
BILL NUMBER: S7722 SPONSOR: OBERACKER TITLE OF BILL: An act to amend the criminal procedure law, in relation to providing that all court clerks of the unified court system shall be designated as peace officers PURPOSE: Would allow Chief Clerks in the 3rd and 4th departments to be allowed Peace Officer status the same as those in the 1st and 2nd departments. SUMMARY OF PROVISIONS: Allows Chief Clerks Peace Officer status in the 3rd and 4th departments of the Unified Court System.