Bill S7745A version original - Iwen Chu
BILL NUMBER: S7745 SPONSOR: CHU TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to requiring batteries for bicycles with electric assist and electric scooters to be tested and approved by and contain the seal of the Underwriters Labora- tories PURPOSE: Requires batteries for bicycles with electric assist and electric scoot- ers to be tested and approved by and contain the seal of the Underwrit- ers Laboratories. SUMMARY OF SPECIFIC PROVISIONS: Section 1.