Bill S7757A version original - Thomas F. O'Mara
BILL NUMBER: S7757 SPONSOR: O'MARA TITLE OF BILL: An act to amend the highway law, in relation to designating certain portions of the state highway system as the "Between Cayuga and Seneca Lakes Wine Trail" PURPOSE: This legislation will create a new wine trail that connects the existing Cayuga and Seneca Lake wine trails SUMMARY OF SPECIFIC PROVISIONS: Section 1. adds subdivision 3-a to section 343-k of the highway law, to create the "between Cayuga and Seneca Lakes Wine Trail" Section 2.