Bill S7751 version original - James Tedisco
BILL NUMBER: S7751 SPONSOR: TEDISCO TITLE OF BILL: An act to amend the education law, in relation to providing additional reimbursement to school districts for expenses incurred as a result of an influx of migrant students entering the schools of the districts PURPOSE: Relates to providing additional reimbursement to school districts for expenses incurred to a student population spike, while providing relief to localities. SUMMARY OF PROVISIONS: § 1.