Bill S7761 version original - James Tedisco
BILL NUMBER: S7761 SPONSOR: TEDISCO TITLE OF BILL: An act to amend the education law, in relation to barring a person convicted of a hate crime from attending any state public school of higher education PURPOSE: Relates to barring a person convicted of a hate crime under Section 485.05 of the New York State Penal Law from attending any state public school of high education. SUMMARY OF PROVISIONS: § 1. The education law is amended by adding a new section 6436-a to read as follows: § 6436-a.