Bill S7759A version original - Michael Gianaris
BILL NUMBER: S7759 SPONSOR: GIANARIS TITLE OF BILL: An act to amend the executive law, in relation to enacting the "keep police radio public act" PURPOSE: This bill ensures that duly authorized members of the press, and the public, maintain access to law enforcement radio communications. SUMMARY OF PROVISIONS: Section one is the title.