Bill S7775 version original - José M. Serrano
BILL NUMBER: S7775 SPONSOR: SERRANO TITLE OF BILL: An act to create a temporary state commission to study and make recom- mendations concerning the application of pesticides in cities with a population of one million or more PURPOSE: To protect the residents of New York State from the possible health hazard posed. by the repeated spreading of pesticides. SUMMARY OF PROVISIONS: Directs the creation of a temporary state commission to study the appli- cation of pesticides in.