Bill S7715A version A - Jeremy Cooney
BILL NUMBER: S7715A SPONSOR: COONEY TITLE OF BILL: An act to amend the criminal procedure law, in relation to criminal possession of stolen property SUMMARY OF SPECIFIC PROVISIONS: Section 1 amends section 510.10 of the criminal procedure law by adding the crimes of criminal possession of stolen property in the fourth, third, second, and first, as well as grand larceny in the fourth, third, and second to the list of qualifying offenses. Section 2 amends section 530.