Bill S7795 version original - Julia Salazar
BILL NUMBER: S7795 SPONSOR: SALAZAR TITLE OF BILL: An act to repeal certain provisions of the correction law relating to the transfer of incarcerated individuals from a county jail to a state correctional institution PURPOSE OR GENERAL IDEA OF BILL:: This bill repeals paragraph 2 of Section 504 of the Correction Law. Repealing this section would continue to permit jail-to-jail transfers in accordance with the law but would eliminate the current provision that permits jail to prison transfers.