Bill S7820 version original - Shelley Mayer
BILL NUMBER: S7820 SPONSOR: MAYER TITLE OF BILL: An act authorizing the Chai Center to file an application for retroac- tive real property tax exemption PURPOSE OR GENERAL IDEA OF BILL: To allow the Chai Center to file an application for retroactive real property tax exemption for the 2021-2022 assessment roll. SUMMARY OF PROVISIONS: Section 1 gives authorization to the assessor of the town of Eastchester to accept an application for exemption from The Chai Center.