Bill S7827 version original - Jessica Ramos
BILL NUMBER: S7827 SPONSOR: RAMOS TITLE OF BILL: An act to amend the judiciary law, in relation to requiring the report- ing of completed judicial training as mandated by part 17 of the rules of the chief judge of the New York state unified court system PURPOSE: To ensure compliance with mandatory judicial training as outlined by the rules of the chief judge of the New York state unified court system. SUMMARY OF PROVISIONS: Section 1.