Bill S2146A version A - Andrew Gounardes
BILL NUMBER: S2146A SPONSOR: GOUNARDES TITLE OF BILL: An act to amend the education law and the state finance law, in relation to enacting The New Deal for CUNY PURPOSE OR GENERAL IDEA OF BILL: This bill is intended to reestablish the City University of New York as an engine for social and economic mobility by waiving all tuition and creating certain staff-to-student ratios in order to Increase student success. SUMMARY OF PROVISIONS: Section 1 names the title of the bill.