Bill S7882 version original - Monica Martinez
BILL NUMBER: S7882 SPONSOR: MARTINEZ TITLE OF BILL: An act to amend the public health law, in relation to prohibiting the sale of vape products that resemble school supplies PURPOSE: The purpose of this legislation is to ensure any place of business where tobacco and other vaping products are sold do not carry products inten- tionally packaged to appeal to our youth population in the form of school supplies and other related products. SUMMARY OF PROVISIONS: Section 1.