Bill S7938 version original - Pete Harckham
BILL NUMBER: S7938 SPONSOR: HARCKHAM TITLE OF BILL: An act directing the commissioner of the division of criminal justice services to study gender disparity in sentencing outcomes PURPOSE: Directs the commissioner of the division of criminal justice services to study gender disparity in sentencing outcomes. SUMMARY OF PROVISIONS: Section 1 directs the commissioner of the Department of Corrections and Community Supervision to study gender disparity in sentencing.