Bill S7911 version original - Kevin S. Parker
BILL NUMBER: S7911 SPONSOR: PARKER TITLE OF BILL: An act to amend the executive law and the state finance law, in relation to creating the New York state autism council PURPOSE OR GENERAL IDEA OF BILL: This bill authorizes the creation of the New York State Autism Council. SUMMARY OF SPECIFIC PROVISIONS: Section 1 of the bill defines the terms used within the legislation and sets forth the makeup of the council as well as the responsibilities and duties of the council.