Bill S7877 version original - Mark Walczyk
BILL NUMBER: S7877 SPONSOR: WALCZYK TITLE OF BILL: An act to amend the criminal procedure law, in relation to including federal law enforcement officers in the definition of police officer PURPOSE: This bill would include federal police in the definition of a police officer. SUMMARY OF PROVISIONS: Section 1: Amends subdivision 34 of section 1.20 of the criminal proce- dure law. Section 2: Sets the effective date.