Bill S7899 version original - Kevin S. Parker
BILL NUMBER: S7899 SPONSOR: PARKER TITLE OF BILL: An act to amend the general construction law, in relation to the desig- nation of Kwanzaa as a public holiday PURPOSE OR GENERAL IDEA OF BILL: The purpose of the legislation is to establishes Kwanzaa as a public holiday. SUMMARY OF SPECIFIC PROVISIONS: Between December 26th and the first day of January shall be known as Kwanzaa. PRIOR LEGISLATIVE HISTORY: 2021-22: S6830 - Referred to Judiciary FISCAL IMPLICATIONS: None.