Bill S7903 version original - Kevin S. Parker
BILL NUMBER: S7903 SPONSOR: PARKER TITLE OF BILL: An act to amend the judiciary law, in relation to optional excusal from jury service for persons age 70 and over PURPOSE OR GENERAL IDEA OF BILL: Allows persons to opt out of state jury service based solely on age without the burden of traveling to the courthouse to make the applica- tion similar to the federal court system. SUMMARY OF PROVISIONS: Section 1 Amends subdivision (a) of section 517 judiciary law by adding a new paragraph 3.