Bill S7922 version original - Nathalia Fernandez
BILL NUMBER: S7922 SPONSOR: FERNANDEZ TITLE OF BILL: An act to amend the general business law, in relation to requiring publishers of books created with the use of generative artificial intel- ligence to contain a disclosure of such use PURPOSE OR GENERAL IDEA OF BILL: This bill would require publishers. to disclose whether any portion of a book was published with the use of generative artificial intelligence software to provide transparency to consumers making purchasing deci- sions.