Bill S7930 version original - Anthony H. Palumbo
BILL NUMBER: S7930 SPONSOR: PALUMBO TITLE OF BILL: An act to amend chapter 234 of the laws of 1907 relating to authorizing the formation of a park district in the town of Southold, Suffolk coun- ty, to acquire lands for park purposes, and to issue bonds therefor, in relation to appointment of a treasurer by the park commissioners PURPOSE: To provide for the appointment of the treasurer of the Southold Park District. SUMMARY OF PROVISIONS: Section 1.