Bill S7867 version original - Rachel May
BILL NUMBER: S7867 SPONSOR: MAY TITLE OF BILL: An act to amend the transportation law, in relation to requiring the department of transportation to develop an updated New York state bicy- cle and pedestrian plan PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to establish regular updates to the state's plan for bicycle and pedestrian transportation infrastructure. SUMMARY OF PROVISIONS: Section 1 requires the department to update its plan and to continue to do so every five years.