Bill S7845 version original - Neil D. Breslin
BILL NUMBER: S7845 SPONSOR: BRESLIN TITLE OF BILL: An act to amend the insurance law, in relation to providing for the issuance of pet insurance PURPOSE: The purpose of this bill is to regulate and define the issuance of pet insurance. SUMMARY OF PROVISIONS: Section 1 - Amends subsection (a) to section 1113 of the insurance law to include a new paragraph including "Pet insurance". Section 2 - Amends the insurance law by adding a new section 3462. Section 3 - Establishes the effective date.