Bill S7847 version original - Lea Webb
BILL NUMBER: S7847 SPONSOR: WEBB TITLE OF BILL: An act to amend the general business law, in relation to requiring publications to identify when the use of artificial intelligence is present within such publication PURPOSE: The purpose of this bill is to require publishers of newspapers, maga- zines or other periodical publications to disclose whether any portion of such publication was created with the use of generative artificial intelligence. SUMMARY OF PROVISIONS: Section 1.