Bill S7881 version original - Monica Martinez
BILL NUMBER: S7881 SPONSOR: MARTINEZ TITLE OF BILL: An act to amend the penal law, in relation to a threat to share an inti- mate image PURPOSE: The purpose of this bill is to classify the threat to disseminate or publish an intimate image or video as a criminal offense. SUMMARY OF PROVISIONS: Section 1. Subdivision 1 of section 245.