Bill S7862 version original - Steve Rhoads
BILL NUMBER: S7862 SPONSOR: RHOADS TITLE OF BILL: An act to amend the penal law, in relation to allowing members of the Nassau County auxiliary police force to possess a police baton when authorized by the police commissioner of such county SUMMARY OF PROVISIONS: Section 1 of the bill amends subdivision b of section 265.20 of the penal law, by adding "or the county of Nassau". Section 2 of the bill provides the effective date.