Bill S7939 version original - Pete Harckham
BILL NUMBER: S7939 SPONSOR: HARCKHAM TITLE OF BILL: An act to amend the tax law, in relation to authorizing the town of Cortlandt to impose a hotel and motel tax; and providing for the repeal of such provisions upon expiration thereof PURPOSE: Authorizes the town of Cortlandt to impose a hotel and motel tax. SUMMARY OF PROVISIONS: Section 1 amends the tax law to authorize the town of Cortlandt to impose a hotel and motel tax.