Bill S7864A version original - Michelle Hinchey
BILL NUMBER: S7864 SPONSOR: HINCHEY TITLE OF BILL: An act to amend the alcoholic beverage control law, in relation to the sale of non-alcoholic versions of alcoholic beverages PURPOSE: To allow liquor stores to sell non-alcoholic versions of alcoholic beverages. SUMMARY OF PROVISIONS: Section one amends section 3 of the alcoholic beverage control law to create a new definition of non-alcoholic version of an alcoholic bever- age.