Bill S7860 version original - Liz Krueger
BILL NUMBER: S7860 SPONSOR: KRUEGER PURPOSE: This bill would require that moped dealers be registered appropriately with the Department of Motor Vehicles to be allowed to offer mopeds for sale. SUMMARY OF SPECIFIC PROVISIONS: Section one increases the penalties for unlicensed dealers who sell mopeds without the appropriate registration with the State.