Bill S7866 version original - Rachel May
BILL NUMBER: S7866 SPONSOR: MAY TITLE OF BILL: An act to amend the general business law, in relation to journalism usage fees PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to create a process where local news publishers are compensated for their work used on social media plat- forms. SUMMARY OF PROVISIONS: Section 1.