Bill S7976 version original - Jessica Ramos
BILL NUMBER: S7976 SPONSOR: RAMOS TITLE OF BILL: An act to amend the labor law, in relation to local enforcement of violations of the state minimum wage in a city with a population of one million or more PURPOSE: To establish local enforcement authority for the Comptroller of the City of New York, who shall be empowered to enforce and investigate violations of State minimum wage law. SUMMARY OF PROVISIONS: Section 1.