Bill S7994 version original - Kevin S. Parker
BILL NUMBER: S7994 SPONSOR: PARKER TITLE OF BILL: An act to amend the public service law, in relation to modifying the definition of renewable energy systems PURPOSE OR GENERAL IDEA OF BILL: To allow for the continued operation of biomass facilities located on a military installation and providing 100% of the military base's elec- tricity. SUMMARY OF PROVISIONS: Section 1: Includes "biomass within a military installation" within the definition of renewable energy systems.