Bill S7962 version original - Kevin S. Parker
BILL NUMBER: S7962 SPONSOR: PARKER TITLE OF BILL: An act to amend the public health law, in relation to the definition of elevated blood lead levels PURPOSE OR GENERAL IDEA OF BILL: This bill will amend the numeric definition of elevated lead levels in public health law. SUMMARY OF PROVISIONS: Section 1 - amends subdivision 6 of section 1370 of the public health law, to redefine the definition of "elevated lead levels" to 3.5 micro- grams.