Bill S7981 version original - Cordell Cleare
BILL NUMBER: S7981 SPONSOR: CLEARE TITLE OF BILL: An act to amend the correction law, in relation to creating parity for incarcerated individual telephone services for international calls to family members PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to create parity and ensure access for incarcerated :Individuals telephone services when making international calls to other countries for example countries such as Ecuador, Israel, and Philippines. SUMMARY OF PROVISIONS: Section 1.