Bill S7984 version original - Cordell Cleare
BILL NUMBER: S7984 SPONSOR: CLEARE TITLE OF BILL: An act to amend the cannabis law and the public health law, in relation to limiting the hours of certain businesses located near schools PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to deter the youth from entering adult-use facilities and tobacco businesses. Additionally, it aims to hinder the establishments' ability to promote their products to minors.