Bill S7942 version original - Brad Hoylman-Sigal
BILL NUMBER: S7942 SPONSOR: HOYLMAN-SIGAL TITLE OF BILL: An act to amend the penal law, in relation to the age of persons eligi- ble for a sentence of life imprisonment without parole PURPOSE: To ban the imposition of life-without-parole sentences for crimes committed as a minor. SUMMARY OF SPECIFIC PROVISIONS: Section 1 amends subdivision 5 of section 70.