Bill S7945 version original - William Weber
BILL NUMBER: S7945 SPONSOR: WEBER TITLE OF BILL: An act to amend the penal law, in relation to prohibiting falsely reporting an incident and/or making a terroristic threat which include claims of a deadly weapon or dangerous instrument being involved or which threaten a school or place of worship PURPOSE: To strengthen the laws that address swatting. SUMMARY OF PROVISIONS: Section 1. Section 240.