Bill S7957 version original - Kevin S. Parker
BILL NUMBER: S7957 SPONSOR: PARKER TITLE OF BILL: An act to amend the executive law, in relation to creating the Brooklyn health care commission PURPOSE: To create a commission to evaluate the health care network in Kings County, and to make recommendations for solutions to the healthcare crisis in Kings County. SUMMARY OF PROVISIONS: Section One a of the bill amends executive law to add Article 49-C in order to create the Brooklyn health care commission.