Bill S7954 version original - Kevin S. Parker
BILL NUMBER: S7954 SPONSOR: PARKER TITLE OF BILL: An act to amend the education law, in relation to enacting the "free resources for equitable education (FREE) act" PURPOSE: The purpose of this bill is,to lower one of the largest ancillary costs associated with pursuing higher education, while ensuring that assigned texts paid for are a necessity to students' education. SUMMARY OF SPECIFIC PROVISIONS: Section 1 of the bill names the bill the "free resources for equitable education (FREE) act".