Bill S7959 version original - Kevin S. Parker
BILL NUMBER: S7959 SPONSOR: PARKER TITLE OF BILL: An act to amend the mental hygiene law, in relation to requiring video cameras in facilities operated by the office for people with develop- mental disabilities PURPOSE OR GENERAL IDEA OF BILL: To allow the Office for People with Developmental Disabilities to place video cameras on the entrance and exit of all facilities. SUMMARY OF SPECIFIC PROVISIONS: Section 1.