Bill S7946 version original - Kevin S. Parker
BILL NUMBER: S7946 SPONSOR: PARKER TITLE OF BILL: An act to amend the labor law and the civil service law, in relation to employees' right to review personnel records PURPOSE OR GENERAL IDEA OF BILL: This bill would amend the labor and civil service laws to provide a means by which public and private employees are granted access to their personnel file. SUMMARY OF SPECIFIC PROVISIONS: Section 1 of this bill amends the labor law by adding section 211-b.