Bill S7977A version original - Rachel May
BILL NUMBER: S7977 SPONSOR: MAY TITLE OF BILL: An act to amend the environmental conservation law, in relation to banning the use of polystyrene foam docks, buoys and floating structures PURPOSE: The purpose of this act is to prohibit the use of expanded polystyrene foam in buoys, docks, or floating structures on state waters unless the foam is encapsulated or designed to prevent disintegration into the water. Open-cell (beaded) polystyrene is also prohibited.