Bill S6777A version A - Cordell Cleare
BILL NUMBER: S6777A SPONSOR: CLEARE TITLE OF BILL: An act to amend the general business law, in relation to authorizing owners of mobile barber shops to obtain barber shop owner's licenses PURPOSE OR GENERAL IDEA OF BILL: Removes the prohibition against itinerant barber shops and allows for mobile barber shops to be licensed in the State. SUMMARY OF PROVISIONS: Section 1: Amends section 430 of the General Business Law to remove language related to the itinerant barbers.