Bill S8063 version original - Cordell Cleare
BILL NUMBER: S8063 SPONSOR: CLEARE TITLE OF BILL: An act to amend the judiciary law, in relation to hope cards and to amend a chapter of the laws of 2023 amending the judiciary law relating to enacting the "hope card act", as proposed in legislative bills numbers S. 5935 and A. 5040, in relation to the effectiveness thereof PURPOSE: This is a chapter amendment to L.2023, ch. 538.