Bill S8074 version original - Liz Krueger
BILL NUMBER: S8074 SPONSOR: KRUEGER TITLE OF BILL: An act to amend the general construction law, in relation to requiring the use of gender-neutral terms in law PURPOSE: This is a chapter amendment to L.2022, ch.139. Most notably, it modifies the types of legislation that must be interpreted or drafted in a gender-neutral manner. SUMMARY OF PROVISIONS: Section 1 makes three changes to the underlying chapter.