Bill S8058 version original - Brad Hoylman-Sigal
BILL NUMBER: S8058 SPONSOR: HOYLMAN-SIGAL TITLE OF BILL: An act to amend the family court act, the executive law, the civil prac- tice law and rules, the criminal procedure law, the public health law, the education law and the insurance law, in relation to gender-affirming care PURPOSE: This is a chapter amendment to L.2022, ch.143, to expand protections for gender-affirming care. SUMMARY OF PROVISIONS: Section 1 makes three substantive changes to the underlying chapter.