Bill S8044 version original - Zellnor Myrie
BILL NUMBER: S8044 SPONSOR: MYRIE TITLE OF BILL: An act to amend the executive law, in relation to defining the term "mass shooting" for purposes of emergency response measures and access to emergency funding PURPOSE: This is a chapter amendment that makes changes to the executive law. SUMMARY OF PROVISIONS: This amendment makes changes to the definition of mass shooting by removing language that includes individuals who are injured in a shoot- ing incident.