Bill S8067 version original - James Skoufis
BILL NUMBER: S8067 SPONSOR: SKOUFIS TITLE OF BILL: An act to amend the parks, recreation and historic preservation law, in relation to the term veteran and Gold Star families and the free use of campsites PURPOSE OR GENERAL IDEA OF BILL: This is a chapter amendment that makes changes to provisions of Chapter 616 of the laws of 2023 to revise certain definitions. SUMMARY OF PROVISIONS: Section 1 updates the terms "veterans" and "member of a Gold Star fami- ly." Section 2 sets the effective date.