Bill S8041 version original - José M. Serrano
BILL NUMBER: S8041 SPONSOR: SERRANO TITLE OF BILL: An act to amend the parks, recreation and historic preservation law, in relation to littering and dumping on state park lands; and to repeal certain provisions of such law relating thereto PURPOSE OR GENERAL IDEA OF BILL: This is a chapter amendment that makes changes to provisions of Chapter 640 of the laws of 2023 to revise certain definitions. SUMMARY OF PROVISIONS: Section 1 repeals Section 3.