Bill S8062 version original - Cordell Cleare
BILL NUMBER: S8062 SPONSOR: CLEARE TITLE OF BILL: An act to amend the state technology law, in relation to providing for the remote conduct of witnessing health care proxies PURPOSE: This is a chapter amendment that makes changes to provisions of L.2023, c.636, to allow remote witnessing of healthcare proxies. SUMMARY OF PROVISIONS: This legislation makes conforming changes to the State Technology Law. JUSTIFICATION: This legislation is a negotiated change to the underlying chapter.