Bill S8132 version original - Kevin S. Parker
BILL NUMBER: S8132 SPONSOR: PARKER TITLE OF BILL: An act to amend the executive law, in relation to establishing the office of ethnic and community media SUMMARY OF PROVISIONS: Section 1: Amends the executive law by adding a new section 170-h. Establishes the office of ethnic and community media and describes the duties of its executive director who shall be appointed by the governor.