Bill S8098 version original - Andrew Gounardes
BILL NUMBER: S8098 SPONSOR: GOUNARDES TITLE OF BILL: An act to amend the education law, in relation to the practice of dentistry PURPOSE OR GENERAL IDEA OF BILL: To eliminate unnecessary language in the definition of dental practice. SUMMARY OF PROVISIONS: Section one of the bill deletes superfluous language that refers to restoring and maintaining dental health. Section two of this bill provides that this act shall take effect imme- diately.