Bill S8111 version original - Monica Martinez
BILL NUMBER: S8111 SPONSOR: MARTINEZ TITLE OF BILL: An act to amend the veterans' services law, in relation to establishing the "Vets to Vollies" program PURPOSE OR GENERAL IDEA OF BILL: Establishes the "Vets to Vollies" program. SUMMARY OF PROVISIONS: 1 Section 1. The veterans' services law is amended by adding a new section 29-b to read as follows: § _29-b. Vets to vollies. 1.