Bill S8133 version original - Kevin S. Parker
BILL NUMBER: S8133 SPONSOR: PARKER TITLE OF BILL: An act to amend the mental hygiene law, in relation to requiring peti- tioners for appointment of a guardian to identify other persons who may be able to manage the affairs of an incapacitated person PURPOSE OF BILL: This legislation would prevent the use of the guardianship law for the primary purpose of bill collection and would correct erroneous cross. references to the power of attorney law.