Bill S932D version C - Roxanne J. Persaud
BILL NUMBER: S932C SPONSOR: PERSAUD TITLE OF BILL: An act to amend the general business law, in relation to the cancella- tion of a health club contract PURPOSE: To enhance consumer protections surrounding health club contracts by expanding the parameters and methods by which such contracts may be cancelled. SUMMARY OF PROVISIONS: Section 1 of the bill amends subdivision 2 and 3 of section 624 of the general business law and adds a new subdivision 4.