Bill S8145 version original - Timothy M. Kennedy
BILL NUMBER: S8145 SPONSOR: KENNEDY TITLE OF BILL: An act to amend the local finance law, in relation to the sale of munic- ipal obligations by the county of Erie PURPOSE OR GENERAL IDEA OF BILL: To provide Erie County with sufficient flexibility in its bond sales to maximize its return on bonds sold. SUMMARY OF PROVISIONS: Section 1. Amends section 54.