Bill S8151 version original - Joseph P. Addabbo Jr.
BILL NUMBER: S8151 SPONSOR: ADDABBO TITLE OF BILL: An act to amend the cannabis law, in relation to prohibiting cannabis retail and on-site consumption licensees from operating such premises within one thousand feet of the perimeter of a school grounds PURPOSE OF BILL: The purpose of this bill all measurements for determining the location of a cannabis retail licensees are based off the perimeter of a school grounds or house of worship. SUMMARY OF PROVISIONS: Section 1.