Bill S5597B version A - Leroy Comrie
BILL NUMBER: S5597A SPONSOR: COMRIE TITLE OF BILL: An act to amend the banking law, in relation to what mortgage applicants need to know pamphlet on residential mortgages PURPOSE OR GENERAL IDEA OF BILL: This bill would require the creation and publication of the "what mort- gage applicants need to know" pamphlet. The pamphlet shall be provided to consumers by mortgage lenders and bankers, and shall be published in the six most spoken languages in New York State.