Bill S8154 version original - Jessica Ramos
BILL NUMBER: S8154 SPONSOR: RAMOS TITLE OF BILL: An act to amend the labor law, in relation to the minimum wage outside of New York city and the counties of Nassau, Suffolk, and Westchester; and to repeal certain provisions of such law relating to exceptions to minimum wage increases PURPOSE: Amends labor law to establish minimum wage parity across the state. SUMMARY OF PROVISIONS: Section 1.