Bill S8164 version original - James Skoufis
BILL NUMBER: S8164 SPONSOR: SKOUFIS TITLE OF BILL: An act to amend the tax law, in relation to authorizing the town of Chester to impose a hotel and motel tax; and providing for the repeal of such provisions upon expiration thereof PURPOSE: Allows the town of Chester to levy a hotel and motel tax, increasing revenue for the town's general fund. SUMMARY OF PROVISIONS: Section 1: Authorizes the Town of Chester to adopt and amend local laws imposing a tax on hotel or motel rooms.