Bill S8181 version original - Robert Rolison
BILL NUMBER: S8181 SPONSOR: ROLISON TITLE OF BILL: An act to amend the penal law, in relation to penalties for promotion of a suicide attempt PURPOSE OF THE BILL: To increase penalties for endangering the welfare of an incompetent or physically disabled person or a vulnerable elderly person when it significantly contributes to a victim's suicide attempt. SUMMARY OF SPECIFIC PROVISIONS: Section 1 of this bill amends section 120.