Bill S8191 version original - Pete Harckham
BILL NUMBER: S8191 SPONSOR: HARCKHAM TITLE OF BILL: An act to amend the tax law, in relation to establishing a tax credit for the adoption of pets PURPOSE OR GENERAL IDEA OF BILL: Provides for tax credit for the adoption of pets from animal shelters up to $250. SUMMARY OF PROVISIONS: Section 1 amends Section 606 of the tax law by adding a new subsection (kkk) allowing for a tax credit for the adoption of pets. Section 2 sets the effective date.