Bill S8180 version original - Robert Rolison
BILL NUMBER: S8180 SPONSOR: ROLISON TITLE OF BILL: An act to amend the executive law, in relation to eligibility for awards for victims of a crime PURPOSE OF THE BILL: To allow the families of victims of crimes who subsequently commit suicide to receive monetary awards from the Office of Victim Services (OVS). SUMMARY OF SPECIFIC PROVISIONS: Section 1 amends Executive Law § 624, subdivision 1 by correcting a grammatical error.