Bill S8192 version original - John Mannion
BILL NUMBER: S8192 Revised 5/10/24 SPONSOR: MANNION TITLE OF BILL: An act to amend the public health law and the education law, in relation to requiring prescriptions to include a quick response code containing certain information PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to authorize the use of quick response codes to interpret labels affixed to prescription drugs. Section 1 of this bill would define relative terms.