Bill S8195 version original - Leroy Comrie
BILL NUMBER: S8195 SPONSOR: COMRIE TITLE OF BILL: An act to amend the election law, in relation to enacting the "disaster voting relief act of 2024" PURPOSE: This bill would ensure that registered voters are given an opportunity to cast their votes when a natural or manmade disaster has affected their polling sites. SUMMARY OF PROVISIONS: Section one of this bill titles the bill the "Disaster Voting Relief_ Act of 2021.