Bill S8210 version original - Jeremy Cooney
BILL NUMBER: S8210 SPONSOR: COONEY TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to clarifying that all trail users, including bicyclists, have right of way in cross- walks wherever the multi-use trail crosses a roadway PURPOSE:. Bill to clarify that bicyclists on multi-use trails have the right-of- way when crossing a roadway within a crosswalk. SUMMARY OF PROVISIONS: Section 1. Amends 1151 of the vehicle and traffic law by adding a new subdivision (d).