Bill S8227 version original - George Borrello
BILL NUMBER: S8227 SPONSOR: BORRELLO TITLE OF BILL: An act in relation to authorizing the city of Olean assessor to accept an application for a real property tax exemption from Olean Community Theatre, Inc. PURPOSE OR GENERAL IDEA OF BILL: This bill authorizes the city of Olean assessor to accept an application for a real property tax exemption from Olean Community Theatre, Inc. with respect to the 2021-2022 assessment roll.