Bill S8330 version original - Monica Martinez
BILL NUMBER: S8330 SPONSOR: MARTINEZ TITLE OF BILL: An act in relation to authorizing the assessor of the town of Babylon to accept an application for exemption from real property taxes from the Phoenix Houses of Long Island, Inc. PURPOSE: The purpose of this bill is to allow the town of Babylon to accept an application for a retro-active property tax exemption from the Phoenix Houses of Long Island, Inc. SUMMARY OF PROVISIONS: Section 1.