Bill S8333 version original - Jessica Ramos
BILL NUMBER: S8333 SPONSOR: RAMOS TITLE OF BILL: An act to amend the civil service law, in relation to health insurance for state officers and employees who receive health insurance benefits from the state PURPOSE OR GENERAL IDEA OF BILL: The purpose of this legislation is to allow the health benefit plan for new state officers and employees to take place immediately upon employ- ment and not after a 28-day waiting period. SUMMARY OF PROVISIONS: Section 1.