Bill S8337 version original - George Borrello
BILL NUMBER: S8337 SPONSOR: BORRELLO TITLE OF BILL: An act to amend the penal law, in relation to increasing the maximum sentence of imprisonment for certain misdemeanors to one year; and to repeal certain provisions of such law relating thereto PURPOSE OR GENERAL IDEA OF BILL: This bill increases the maximum penalty for class A misdemeanors to one year. SUMMARY OF PROVISIONS: Section 1 repeals subdivision 1-a of section 70.