Bill S8322 version original - Leroy Comrie
BILL NUMBER: S8322 SPONSOR: COMRIE TITLE OF BILL: An act to amend the general business law, in relation to prohibiting the sale of covered license plate frames PURPOSE OR GENERAL IDEA OF BILL: License plate covers are illegal in New York State and yet are still available for sale. This bill prohibits their sale, aligning the sale of the product with the law.