Bill S8338A version original - George Borrello
BILL NUMBER: S8338 SPONSOR: BORRELLO TITLE OF BILL: An act to amend the highway law, in relation to designating a portion of the state highway system as the "T Sgt. Walter Shearing Memorial High- way" PURPOSE OR GENERAL IDEA OF BILL: Designates the Portion of State route 19 from Eddy Road in the Town of Gainesville, county of Wyoming to East Mill Street in the village of Gainesville, county of Wyoming, as the "Tech. Sgt. Walter W.