Bill S8350 version original - Samra Brouk
BILL NUMBER: S8350 SPONSOR: BROUK TITLE OF BILL: An act to amend the education law, in relation to creating the school- based mental health loan repayment program PURPOSE OR GENERAL IDEA OF BILL: To establish a School-based Mental Health Loan Repayment Program for mental health professionals working in schools with ratios not meeting a ratio of students to professionals per the recommendation of the depart- ment of education. SUMMARY OF PROVISIONS: Section 1. Legislative Findings. Section 2.