Bill S8361A version original - Rachel May
BILL NUMBER: S8361 SPONSOR: MAY TITLE OF BILL: An act to amend the state finance law, in relation to requiring all state agencies and offices to cease purchasing single-use plastic water bottles and provides for the installation of water bottle filling stations PURPOSE: To help reduce the state's dependence on single-use plastics and further the state's compliance with the Climate Leadership and Community Protection Act. SUMMARY OF PROVISIONS: Section 1 sets the legislative findings and intent.