Bill S8355 version original - Luis R. Sepúlveda
BILL NUMBER: S8355 SPONSOR: SEPULVEDA TITLE OF BILL: An act to amend the tax law, in relation to making technical changes relating to the rescindment of certificates of authority PURPOSE: Makes technical changes relating to the rescindment of certificates of authority. SUMMARY OF PROVISIONS: Section one: Amends subparagraph A of paragraph 4 of subdivision a of section 1134 of the tax law with respect to only the person's obligation to pay taxes.