Bill S8371 version original - Anthony H. Palumbo
BILL NUMBER: S8371 SPONSOR: PALUMBO TITLE OF BILL: An act to amend the general municipal law, in relation to the sale of motor vehicle tires PURPOSE: This bill would require that the manufacture date of motor vehicle tires be included on the consumer's receipt when purchased. SUMMARY OF PROVISIONS: Section 1. The general business law is amended by adding a new section 391-w to include the receipt of the sale of motor vehicle tires for providing the manufacture date of tires purchased. Section 2.