Bill S2148B version B - Joseph P. Addabbo Jr.
BILL NUMBER: S2148B SPONSOR: ADDABBO TITLE OF BILL: An act to amend the administrative code of the city of New York, in relation to the pensionable earnings of first grade police officers PURPOSE: To provide equitable benefits in the police pension fund for experienced police officers. SUMMARY OF PROVISIONS: Section one of this bill declares that it is necessary to provide this pension benefit in order to enhance public safety and prevent the loss of vital public services in this state.