Bill S222A version A - Zellnor Myrie
BILL NUMBER: S222A SPONSOR: MYRIE TITLE OF BILL: An act to amend the executive law, in relation to establishing applica- tion processing and review requirements for reprieves, commutations and pardons PURPOSE: To amend the Executive Law, in relation to establishing application processing and review requirements for reprieves, commutations and pardons. SUMMARY OFPROVISIONS: Section 1 provides the title of the bill. Section 2 establishes the legislative findings and intent.